requirements.txt
mediapipe==0.10.9
numpy==1.26.4
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
Pillow==11.2.1
pyglet==2.1.6
pygltflib==1.16.4
PyOpenGL==3.1.9
PyOpenGL==3.1.9
pyrr==0.10.3
